+++ src/wallet/rpcwallet.cpp
@@ -33,6 +33,7 @@
 #include <wallet/psbtwallet.h>
 #include <wallet/load.h>
 #include <wallet/rpcwallet.h>
+#include <wallet/scriptpubkeyman.h>
 #include <wallet/wallet.h>
 #include <wallet/walletdb.h>
 #include <wallet/walletutil.h>
